+++ database/migrations/2018_11_26_151843_create_automotoras_table.php
@@ -15,7 +15,7 @@ class CreateAutomotorasTable extends Migration
     {
         Schema::create('automotoras', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('ciudad_id');
+            $table->unsignedBigInteger('ciudad_id');
             $table->string('direccion',127);
             $table->string('nombre',63);
             $table->timestamps();
